+++ core/tabs/applications-setup/browsers/zen-browser.sh
@@ -6,13 +6,21 @@ installZenBrowser() {
     if ! command_exists io.github.zen_browser.zen && ! command_exists zen-browser; then
         printf "%b\n" "${YELLOW}Installing Zen Browser...${RC}"
         case "$PACKAGER" in
-        pacman)
-            "$AUR_HELPER" -S --needed --noconfirm zen-browser-bin
-            ;;
-        *)
-            checkFlatpak
-            "$ESCALATION_TOOL" flatpak install --noninteractive flathub io.github.zen_browser.zen
-            ;;
+            apt-get|nala)
+                if [ "$DTYPE" = "ubuntu" ] && command_exists snap; then
+                    "$ESCALATION_TOOL" snap install zen-browser-snap --candidate
+                else
+                    checkFlatpak
+                    "$ESCALATION_TOOL" flatpak install --noninteractive flathub io.github.zen_browser.zen
+                fi
+                ;;
+            pacman)
+                "$AUR_HELPER" -S --needed --noconfirm zen-browser-bin
+                ;;
+            *)
+                checkFlatpak
+                "$ESCALATION_TOOL" flatpak install --noninteractive flathub io.github.zen_browser.zen
+                ;;
         esac
     else
         printf "%b\n" "${GREEN}Zen Browser is already installed.${RC}"
